assurio/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
PyPDF2
google-generativeai==0.7.2
pydantic
python-multipart
aiofiles
jinja2
cloudinary
pinecone
requests 